Функции поменены местами
